[checkout] move header files, change its identation and landing page
