+++ src/components/header/Header.test.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import '../tests/matchMedia'
+import '../../tests/matchMedia'
 import { render } from '@testing-library/react'
 import Header from './Header'
 import { Provider } from 'react-redux'
