+++ src/routes/cart/Cart.jsx
@@ -3,7 +3,7 @@ import useSWR from 'swr'
 import URL from '../../utils/constants.js'
 import { Col, Row, Spin } from 'antd'
 import { useSelector, useDispatch } from 'react-redux'
-import Header from '../../components/Header.jsx'
+import Header from '../../components/header/Header.jsx'
 import { setCartInfo } from '../../reducer/Cart'
 import productContent from './Product'
 import './Cart.css'
