+++ src/routes/payment/Payment.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Col, Row } from 'antd'
-import Header from '../../components/Header'
+import Header from '../../components/header/Header'
 import './Payment.css'
 
 const Payment = () => {
